add new Conversation actions: AddConversation, AddPlainComment, EditPlainComment, DeletePlainComment. add CommentDataHandler and SegmentDataHandler. rename Models/Conversation to Models/Conversations, add Requests and Responses folders inside. add missing tests
