--- Apps.Memsource/Models/Conversation/EditConversationRequest.cs
@@ -1,11 +1,0 @@
-﻿using Apps.PhraseTMS.DataSourceHandlers.StaticHandlers;
-using Blackbird.Applications.Sdk.Common.Dictionaries;
-
-namespace Apps.PhraseTMS.Models.Conversation
-{
-    public class EditConversationRequest
-    {
-        [StaticDataSource(typeof(ConversationStatusDataHandler))]
-        public string? Status { get; set; }
-    }
-}
