--- Apps.Memsource/Models/Conversation/SearchConversationRequest.cs
@@ -1,12 +1,0 @@
-﻿using Blackbird.Applications.Sdk.Common;
-
-namespace Apps.PhraseTMS.Models.Conversation
-{
-    public class SearchConversationRequest
-    {
-        [Display("Include deleted")]
-        public bool? IncludeDeleted { get; set; }
-
-        public DateTime? Since { get; set; }
-    }
-}
