+++ Apps.Memsource/Models/Conversations/Requests/AddEditPlainCommentRequest.cs
@@ -1,0 +1,9 @@
+﻿using Newtonsoft.Json;
+
+namespace Apps.PhraseTMS.Models.Conversations.Requests;
+
+public class AddEditPlainCommentRequest
+{
+    [JsonProperty("comment")]
+    public string Text { get; set; }
+}
