+++ Apps.Memsource/Models/Conversations/Requests/Segment.cs
@@ -1,0 +1,7 @@
+﻿namespace Apps.PhraseTMS.Models.Conversations.Requests;
+
+public class Segment
+{
+    public string? Id { get; set; }
+    public string? Source { get; set; }
+}
