+++ Apps.Memsource/Models/Conversations/Responses/CommentsResponse.cs
@@ -1,0 +1,10 @@
+﻿using Newtonsoft.Json;
+using Apps.PhraseTMS.Models.Conversations.Requests;
+
+namespace Apps.PhraseTMS.Models.Conversations.Responses;
+
+public class CommentsResponse
+{
+    [JsonProperty("comments")]
+    public List<Comment> Comments { get; set; }
+}
