+++ Apps.Memsource/Models/Conversations/Responses/SegmentsResponse.cs
@@ -1,0 +1,8 @@
+﻿using Apps.PhraseTMS.Models.Conversations.Requests;
+
+namespace Apps.PhraseTMS.Models.Conversations.Responses;
+
+public class SegmentsResponse
+{
+    public List<Segment> Segments { get; set; }
+}
